[warnings] fix signed/unsigned compare warnings

+++ libfreerdp/core/childsession.c
@@ -256,7 +256,7 @@ static int transport_bio_named_puts(BIO* bio, const char* str)
 	WINPR_ASSERT(bio);
 	WINPR_ASSERT(str);
 
-	const int max = (INT_MAX > SIZE_MAX) ? SIZE_MAX : INT_MAX;
+	const size_t max = (INT_MAX > SIZE_MAX) ? SIZE_MAX : INT_MAX;
 	const size_t len = strnlen(str, max);
 	if (len >= max)
 		return -1;
