+++ libfreerdp/crypto/tls.c
@@ -251,7 +251,7 @@ static int bio_rdp_tls_puts(BIO* bio, const char* str)
 	if (!str)
 		return 0;
 
-	const int max = (INT_MAX > SIZE_MAX) ? SIZE_MAX : INT_MAX;
+	const size_t max = (INT_MAX > SIZE_MAX) ? SIZE_MAX : INT_MAX;
 	const size_t size = strnlen(str, max);
 	if (size >= max)
 		return -1;
